pyproject.toml
[project]
name = "kracker-barrel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "bcrypt==4.2.0",
    "cffi==1.17.1",
    "cryptography==43.0.3",
    "git-filter-repo==2.45.0",
    "passlib==1.7.4",
    "pycparser==2.22",
    "pycryptodome==3.21.0",
    "psutil==5.9.8",
    "pyyaml==6.0.3",
    "scrypt==0.8.27",
    "tdqm==0.0.1",
    "tqdm==4.67.0",
]
